OCP-1741: Add change of CONF_DIR to ES_PATH_CONF
In ES 6.x this directive has been renamed

Also we need to change the /etc/default/elasticsearch file before
installing plugins, as the elasticsearch-plugin command now seems to
need ES_PATH_CONF set in this file before it will work.

Finally, elasticsearch_major_version is a string, so we correct the
comparisons in the default template.